model/requirements.txt
flask
numpy
tensorflow
pillow
os